# docs/reference/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Virtual Queue Management System API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Аутентификация пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSessionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Завершение сессии пользователя",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Сессия успешно завершена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "400": {
            "description": "Токен отсутствует",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Сессия не найдена или истекла",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "GUID"
      }
    },
    "schemas": {
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "description": "Логин пользователя"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя"
          }
        },
        "required": ["login", "password"]
      },
      "UserSessionResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "description": "ID сессии"
          },
          "userId": {
            "type": "integer",
            "description": "ID пользователя"
          },
          "login": {
            "type": "string",
            "description": "Логин пользователя"
          },
          "fullName": {
            "type": "string",
            "description": "ФИО пользователя"
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия пользователя"
          },
          "email": {
            "type": "string",
            "description": "Email пользователя"
          },
          "token": {
            "type": "string",
            "description": "Токен аутентификации"
          },
          "tokenExpiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Срок действия токена"
          },
          "roleCodes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Роли пользователя"
          }
        }
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Сообщение о результате"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки"
          }
        }
      }
    }
  }
}